requirements.txt
Pillow==10.3.0
Requests==2.32.2
dataclasses-json==0.6.0
spotipy==2.23.0
